NEW: Added amount to listings

+++ app/controllers/market_controller.rb
@@ -20,26 +20,37 @@ class MarketController < ApplicationController
   end
   
   def buy
-    if params[:id]
+    if params[:id] and params[:amount]
+      amount = params[:amount].to_i
       listing = MarketListing.find(params[:id]) rescue nil
-      if listing and listing.location == current_user.location
+      if listing and listing.location == current_user.location and amount >= 1
+        
+        # Check Amount
+        render json: {'error_message': I18n.t('errors.you_cant_buy_that_much')}, status: 400 and return if amount > listing.amount
         
         # Check Balance
-        render json: {'error_message': I18n.t('errors.you_dont_have_enough_credits')}, status: 400 and return unless current_user.units >= listing.price
+        render json: {'error_message': I18n.t('errors.you_dont_have_enough_credits')}, status: 400 and return unless current_user.units >= listing.price * amount
         
         # If listing is item -> else..
         if listing.item?
-          Item.create(location: current_user.location, user: current_user, loader: listing.loader, equipped: false)
+          amount.times do
+            Item.create(location: current_user.location, user: current_user, loader: listing.loader, equipped: false)
+          end
         else
-          Spaceship.create(location: current_user.location, user: current_user, name: listing.loader, hp: SHIP_VARIABLES[listing.loader]['hp'])
+          amount.times do
+            Spaceship.create(location: current_user.location, user: current_user, name: listing.loader, hp: SHIP_VARIABLES[listing.loader]['hp'])
+          end
         end
         
         # Deduct units
-        current_user.update_columns(units: current_user.units - listing.price)
+        current_user.update_columns(units: current_user.units - listing.price * amount)
         
         # Destroy Listing
-        listing.destroy
-        render json: {}, status: 200 and return
+        new_amount = listing.amount - amount
+        listing.update_columns(amount: new_amount)
+        listing.destroy if new_amount == 0
+        
+        render json: {'new_amount': new_amount}, status: 200 and return
       end
     end
     render json: {}, status: 400
@@ -56,15 +67,16 @@ class MarketController < ApplicationController
   
   def sell
     price = generate_price(params[:loader], params[:type], params[:quantity])
+    quantity = params[:quantity].to_i
     if price != nil
       
       # If type == item -> else..
       if params[:type] == "item"
         # Check if user tries to sell more
-        render json: {'error_message': I18n.t('errors.you_dont_have_enough_of_this')}, status: 400 and return if Item.where(loader: params[:loader], user: current_user, location: current_user.location).count < params[:quantity].to_i
+        render json: {'error_message': I18n.t('errors.you_dont_have_enough_of_this')}, status: 400 and return if Item.where(loader: params[:loader], user: current_user, location: current_user.location).count < quantity
         
         # Destroy items
-        Item.where(loader: params[:loader], user: current_user, location: current_user.location).limit(params[:quantity].to_i).destroy_all
+        Item.where(loader: params[:loader], user: current_user, location: current_user.location).limit(quantity).delete_all
         
       elsif params[:type] == "ship" and params[:id]
       
@@ -92,14 +104,10 @@ class MarketController < ApplicationController
       # Generate Listing
       listing = MarketListing.where(loader: params[:loader], location: current_user.location).first rescue nil
       if listing
-        (params[:quantity].to_i).times do
-          MarketListing.create(loader: params[:loader], listing_type: params[:type], location: current_user.location, price: (listing.price * rand(0.95..1.05)).round)
-        end
+        MarketListing.create(loader: params[:loader], listing_type: params[:type], location: current_user.location, price: (listing.price * rand(0.95..1.05)).round, amount: quantity)
       else
         rabat = rand(0.8..1.2)
-        (params[:quantity].to_i).times do
-          MarketListing.create(loader: params[:loader], listing_type: params[:type], location: current_user.location, price: (get_item_attribute(params[:loader], 'price') * rabat * rand(0.95..1.05)).round)
-        end
+        MarketListing.create(loader: params[:loader], listing_type: params[:type], location: current_user.location, price: (get_item_attribute(params[:loader], 'price') * rabat * rand(0.95..1.05)).round, amount: quantity)
       end
       
       render json: {}, status: 200 and return
